Validate GitHub trigger presence and move callback to xlml/utils/github.py
